cleanup: move convert_tool to Client._make_pydantic_function
