+++ mcp_run/__init__.py
@@ -4,7 +4,7 @@ from .profile import Profile
 from .client import Client
 from .config import ClientConfig
 from .plugin import InstalledPlugin
-from .mcp_protocol import MCPServer, convert_tool
+from .mcp_protocol import MCPServer
 
 __all__ = [
     "Tool",
@@ -19,5 +19,4 @@ __all__ = [
     "ServletSearchResult",
     "ProfileSlug",
     "MCPServer",
-    "convert_tool",
 ]
